// client/src/app/api/model/spanDTO.ts
/**
 * Span API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SpanDTO { 
    id: string;
    children: Array<SpanDTO>;
    name: string;
    startTime: string;
    endTime: string;
    /**
     * Duration in nanoseconds
     */
    duration: number;
}